Phase 2.5: lazy-load the modal-renderer set; generic pages 509→200 KB eager
The Phase-2 core-boot still eagerly imported ~300 KB for the universal verse
modal. Investigation showed most of it was dead: modal.js stored the five
registerModal* tab renderers but never invoked them (its tabs are Verse/
Notes/Connections; verse-study.js imports library.js renderers directly).

- Remove the dead registerModal* bridge from modal.js and core-boot —
  verse-study.js (+interlinear.js) and library.js leave the eager graph.
- Extract mem.js from daily.js (memory-verse localStorage/SRS helpers) so
  the modal's sync Memorize-button calls don't drag the daily page module
  onto every page. daily.js imports mem.js; page review state stays put.
- Inline the sidebar Search button + Ctrl+K/'?' hotkeys into core-boot
  (buildSearchNav, verbatim from search.js buildSearchDOM); the search
  engine now loads only on search/ via new entries/search.js.
- Dynamic-import terms.js/places.js in their existing idle slots (modal
  term-tag registration happens on load; its null-guard covers the gap,
  same as the old term-map fetch window). apocrypha-reader.js loads
  fire-and-forget after wireRefLinks. BibleUI.initOLSection returns the
  import().then() promise (its one caller chains .catch); autoTagPlacesIn
  is a fire-and-forget lazy wrapper (all callers are guarded).
- sw.js: precache mem.js + entries/search.js; APP_CACHE_V v185 → v186.

Eager static-import payload: generic pages 509 → 200 KB (−89% vs the
original 1,800 KB app.js); reader 558 KB; maps 385 KB. Verified by
headless-Chromium functional pass: verse modal opens via ref click with
content, search returns results, John 3:16 renders, library browser
populates, Ctrl+K navigates, window.BibleUI surface exact, zero console
errors. (Found pre-existing: the injected Search button has been dead
since the static-sidebar commit — .version-picker anchor gone; noted in
OVERHAUL.md for Phase 5.)
