+++ assets/js/core-boot.js
@@ -1,72 +1,61 @@
-/* core-boot.js — Shared boot sequence for the per-page entry split (Phase 2).
+/* core-boot.js — Shared boot sequence for the per-page entry split (Phase 2/2.5).
  *
- * Replaces the monolithic app.js: every page now loads a thin entry module
- * from assets/js/entries/ which imports boot() and passes its page-specific
- * init callback. core-boot owns exactly the work app.js ran on EVERY page —
- * theme, PWA, storage migrations, Bible metadata, tooltip/modal/ref wiring,
- * version picker, sidebar Search button, history widget, deferred term/place
- * tagging, and the window.BibleUI public API. Page-feature modules (reader,
- * maps, workshop, …) are NOT imported here; each page's entry imports only
- * what that page uses.
+ * Replaces the monolithic app.js: every page loads a thin entry module from
+ * assets/js/entries/ which imports boot() and passes its page-specific init
+ * callback. core-boot owns exactly the work app.js ran on EVERY page — theme,
+ * PWA, storage migrations, Bible metadata, tooltip/modal/ref wiring, version
+ * picker, sidebar Search button + hotkeys, history widget, deferred term/place
+ * tagging, and the window.BibleUI public API.
  *
- * Modules that ARE imported here ship on every page because the universal
- * verse modal depends on their renderers (word study / topics / confessions /
- * fathers / dictionary tabs) and the memorisation helpers. Making those
- * lazy is a follow-up (dynamic import on first tab click), not Phase 2.
+ * Phase 2.5 — the eager import set is only what the first paint truly needs:
+ * core, storage, tooltip, modal, wire, pwa, mem. Everything else is loaded on
+ * demand via dynamic import():
+ *   - terms.js / places.js  → at requestIdleCallback time (same slots the
+ *     tagging already ran in; the modal's term-tag guard tolerates the gap
+ *     exactly as it tolerated the term-map fetch before)
+ *   - apocrypha-reader.js   → fire-and-forget right after wireRefLinks
+ *     (wireApoRefLinks awaits its own abbrev-map fetch anyway)
+ *   - ol-companion.js / places.js on window.BibleUI → lazy async wrappers;
+ *     the only initOLSection caller (verse-study.js) already chains .catch on
+ *     the returned promise, and autoTagPlacesIn callers are fire-and-forget
+ *   - search.js             → never loaded here; the sidebar Search button and
+ *     hotkeys are inlined below (buildSearchNav), and the search page's own
+ *     entry imports the engine statically
  *
  * boot(pageInit, opts):
  *   pageInit   — optional; runs once Bible metadata + shared UI are ready.
  *   opts.early — run pageInit BEFORE autoTagRefs/wireInlineVerses. Only the
- *                reader entry uses this, preserving app.js's original order
- *                (reader init preceded ref auto-tagging; all other page
- *                inits followed it).
+ *                reader entry uses this, preserving app.js's original order.
  */
 'use strict';
 
 import {
   getVersion, setVersion, metaBooks,
   loadVersions, loadBooks, populateVersionPicker, wireVersionPicker,
-  initTheme, READER_URL, onVersionChange, _fireBibleReady
+  initTheme, READER_URL, SEARCH_URL, onVersionChange, _fireBibleReady
 } from './core.js';
 import { _runStorageMigrations } from './storage.js';
 import { buildTooltipDOM } from './tooltip.js';
 import {
   buildModalDOM, openModal, syncModalVersionPicker,
-  registerModalWordStudy, registerModalTopics, registerModalConfessions,
-  registerModalFathers, registerModalDictionary,
   registerMemHelpers, registerAutoTagTerms,
-  initHistoryWidget
+  _showShortcutsOverlay, initHistoryWidget
 } from './modal.js';
 import {
   wireRefLinks, autoTagRefs, wireInlineVerses, updateInlineVerses
 } from './wire.js';
 import { initPWA } from './pwa.js';
-import { buildSearchDOM } from './search.js';
-
-// ── Modal-tab renderer dependencies (site-wide via the verse modal) ────────
-import { renderModalWordStudy } from './verse-study.js';
-import { _memHas, _memAdd, _memRemove, _memRefreshModalBtn } from './daily.js';
-import { renderModalTopics, renderModalConfessions, renderModalFathers, renderModalDictionary } from './library.js';
-import { runAutoTagTerms, autoTagTerms, getTermMap2 } from './terms.js';
-import { runAutoTagPlaces, autoTagPlacesIn } from './places.js';
-import { initOLSection } from './ol-companion.js';
-import { wireApoRefLinks } from './apocrypha-reader.js';
+import { _memHas, _memAdd, _memRemove, _memRefreshModalBtn } from './mem.js';
 
 // ── Register cross-module callbacks ───────────────────────────────────────
-// These connect the modal to feature-module renderers without creating circular
-// imports. The modal calls these registered functions when a user clicks a tab
-// inside the verse modal (word study, topics, etc.).
-registerModalWordStudy(renderModalWordStudy);
-registerModalTopics(renderModalTopics);
-registerModalConfessions(renderModalConfessions);
-registerModalFathers(renderModalFathers);
-registerModalDictionary(renderModalDictionary);
 // Give the modal access to the memorisation helpers so it can show the
-// "Add to memory" button and react to changes.
+// "Add to memory" button and react to changes. (mem.js is the thin storage
+// module split out of daily.js — the modal needs sync answers, so these
+// stay eager; they cost ~3 KB.)
 registerMemHelpers(_memHas, _memAdd, _memRemove, _memRefreshModalBtn);
-// Give the modal's auto-tag feature access to the term map so theological
-// terms inside verse text are underlined with definitions.
-registerAutoTagTerms(autoTagTerms, getTermMap2);
+// registerAutoTagTerms happens lazily when terms.js loads at idle — until
+// then the modal's term-tag guard (_autoTagTermsFn == null) skips tagging,
+// exactly as it did while the term map was still fetching.
 
 // ── Version-change side effects ───────────────────────────────────────────
 // When the user switches translations, update any inline verse elements and
@@ -86,7 +75,9 @@ onVersionChange(function (id) {
 //   - openModal:       topic-page inline <script> blocks on all 10 topic pages
 //   - openReader:      topic-page inline <script> blocks (pass bookId string)
 //   - autoTagPlacesIn: timeline.js after dynamic re-renders; reader.js on lookup
-//   - initOLSection:   ol-companion.js on verse-study page init
+//                      (lazy: fire-and-forget wrapper importing places.js)
+//   - initOLSection:   verse-study.js word-study sections (lazy: returns the
+//                      import().then(...) promise; the caller chains .catch)
 //   - getVersion/setVersion: version picker in topics/_template inline script
 //   Do not add new window.* state outside this object; use entry callbacks instead.
 // VERIFY: From any topic page DevTools console, confirm `window.BibleUI.openModal`
@@ -96,8 +87,15 @@ window.BibleUI = {
   getVersion:      getVersion,
   setVersion:      setVersion,
   openModal:       openModal,
-  initOLSection:   initOLSection,
-  autoTagPlacesIn: autoTagPlacesIn,  // called by reader.js / timeline.js after dynamic renders
+  initOLSection:   function () {
+    var args = arguments;
+    return import('./ol-companion.js').then(function (m) {
+      return m.initOLSection.apply(null, args);
+    });
+  },
+  autoTagPlacesIn: function (el) {
+    import('./places.js').then(function (m) { m.autoTagPlacesIn(el); });
+  },
   // INTENT: Resolves bookId (e.g. "genesis") to its display name (e.g. "Genesis")
   //   via metaBooks before building the reader URL. Falls back to the raw bookId
   //   string if metaBooks hasn't loaded yet, which produces a URL the reader can
@@ -109,6 +107,46 @@ window.BibleUI = {
     window.location.href = READER_URL + '?ref=' + encodeURIComponent(ref);
   }
 };
+
+// ── buildSearchNav ─────────────────────────────────────────────────────────
+// The sidebar Search button + global hotkeys, moved here from search.js so
+// non-search pages don't load the 60 KB search engine for a nav button.
+// Behavior is verbatim from the old buildSearchDOM (minus its search-page
+// init tail, which entries/search.js now owns).
+function buildSearchNav() {
+  var vp = document.querySelector('.version-picker');
+  if (vp && !document.getElementById('bsw-search-btn')) {
+    var btn = document.createElement('button');
+    btn.id        = 'bsw-search-btn';
+    btn.className = 'bsw-search-btn';
+    btn.setAttribute('aria-label', 'Search verses (Ctrl+K)');
+    btn.textContent = 'Search';
+    vp.parentNode.insertBefore(btn, vp);
+    btn.addEventListener('click', function () {
+      var inp = document.getElementById('bsw-search-input');
+      if (inp) { inp.focus(); inp.select(); } else { window.location.href = SEARCH_URL; }
+    });
+  }
+
+  // Ctrl+K: focus the search input if on the search page, otherwise navigate.
+  document.addEventListener('keydown', function (e) {
+    if ((e.ctrlKey || e.metaKey) && e.key === 'k') {
+      e.preventDefault();
+      var inp = document.getElementById('bsw-search-input');
+      if (inp) { inp.focus(); inp.select(); } else { window.location.href = SEARCH_URL; }
+    }
+  });
+
+  // '?' opens the keyboard shortcuts overlay.
+  // '/' focuses the reader search bar (only on the reader page).
+  document.addEventListener('keydown', function (e) {
+    if (e.key !== '?' && e.key !== '/') return;
+    if (e.key === '/' && !document.getElementById('reader-results')) return;
+    var tag = document.activeElement && document.activeElement.tagName;
+    if (tag === 'INPUT' || tag === 'TEXTAREA' || tag === 'SELECT') return;
+    if (e.key === '?') { e.preventDefault(); _showShortcutsOverlay(); }
+  });
+}
 
 // ── boot ──────────────────────────────────────────────────────────────────
 // Main boot sequence, formerly app.js init(). Runs after DOMContentLoaded (or
@@ -123,7 +161,8 @@ window.BibleUI = {
 //      before step 7, matching the reader's original position in app.js).
 //   7. autoTagRefs + wireInlineVerses — no-ops if there's nothing to tag.
 //   8. _fireBibleReady() — notifies any modules waiting on Bible metadata.
-//   9. Term/place tagging — deferred to idle time so it doesn't block render.
+//   9. Term/place tagging — deferred to idle time; the modules themselves are
+//      dynamic-imported in the same idle slot (Phase 2.5).
 export function boot(pageInit, opts) {
   opts = opts || {};
 
@@ -137,9 +176,12 @@ export function boot(pageInit, opts) {
       buildTooltipDOM();
       buildModalDOM();
       wireRefLinks();        // wire any static [data-ref] elements in the initial HTML
-      wireApoRefLinks();     // second pass: wire apocryphal refs not handled by wireRefLinks
+      // Second pass: wire apocryphal refs not handled by wireRefLinks. Lazy —
+      // wireApoRefLinks awaits its own abbrev-map fetch internally, so arriving
+      // a few ms later via dynamic import is indistinguishable.
+      import('./apocrypha-reader.js').then(function (m) { m.wireApoRefLinks(); });
       wireVersionPicker();
-      buildSearchDOM();      // adds the Search button to the sidebar on every page
+      buildSearchNav();      // adds the Search button to the sidebar on every page
 
       if (pageInit && opts.early) pageInit();
 
@@ -157,10 +199,18 @@ export function boot(pageInit, opts) {
       // registered via onBibleReady() in other modules.
       _fireBibleReady();
 
-      // Term auto-tagging is expensive (scans all text nodes); defer to browser idle time
-      // to avoid blocking first interactive paint. Falls back to setTimeout on browsers
-      // that don't support requestIdleCallback (e.g. older Safari).
-      var _runTag = function () { runAutoTagTerms(); };
+      // Term auto-tagging is expensive (scans all text nodes); defer to browser idle
+      // time so it doesn't block first interactive paint — and (Phase 2.5) load the
+      // module itself in that idle slot too. Registration with the modal happens on
+      // load; until then the modal skips term tagging, same as when the term map
+      // hadn't fetched yet. Falls back to setTimeout on browsers without
+      // requestIdleCallback (e.g. older Safari).
+      var _runTag = function () {
+        import('./terms.js').then(function (m) {
+          registerAutoTagTerms(m.autoTagTerms, m.getTermMap2);
+          m.runAutoTagTerms();
+        });
+      };
       if (typeof requestIdleCallback !== 'undefined') {
         requestIdleCallback(_runTag, { timeout: 3000 });
       } else {
@@ -169,7 +219,9 @@ export function boot(pageInit, opts) {
 
       /* Place-name auto-tagging — deferred to a second idle slot so it doesn't
          compete with term tagging for the same idle window.                   */
-      var _runPlaces = function () { runAutoTagPlaces(); };
+      var _runPlaces = function () {
+        import('./places.js').then(function (m) { m.runAutoTagPlaces(); });
+      };
       if (typeof requestIdleCallback !== 'undefined') {
         requestIdleCallback(_runPlaces, { timeout: 5000 });
       } else {
