+++ src/pages/search/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Explore — Kingdom Bible Study"} description={"Explore the Bible — search by keyword or phrase, browse topics, study guides, dictionary, and word cloud."}>
+<Base entry="search" title={"Explore — Kingdom Bible Study"} description={"Explore the Bible — search by keyword or phrase, browse topics, study guides, dictionary, and word cloud."}>
 <Fragment slot="head-end">
 <style is:inline="">
     /* ── Explore hub iframe panels ───────────────────────────── */
